Refactored Code for Usage in Manager Tab
