+++ app/femr/ui/controllers/manager/ManagerController.java
@@ -1,4 +1,4 @@
-package femr.ui.controllers.admin;
+package femr.ui.controllers.manager;
 
 import com.google.inject.Inject;
 import femr.business.services.core.IEncounterService;
@@ -12,18 +12,15 @@ import femr.common.models.PatientItem;
 import femr.data.models.mysql.Roles;
 import femr.ui.helpers.security.AllowedRoles;
 import femr.ui.helpers.security.FEMRAuthenticated;
+import femr.ui.models.manager.IndexViewModelGet;
 import femr.ui.models.sessions.CreateViewModel;
-import femr.ui.models.triage.ManageViewModelPost;
-import femr.ui.views.html.admin.triagePatients.triagePatient;
-import femr.util.calculations.dateUtils;
-import org.joda.time.DateTime;
+import femr.ui.views.html.manager.index;
 import play.data.Form;
 import play.mvc.Controller;
 import play.mvc.Result;
 import play.mvc.Security;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 /**
@@ -31,8 +28,8 @@ import java.util.List;
  */
 
 @Security.Authenticated(FEMRAuthenticated.class)
-@AllowedRoles({Roles.PHYSICIAN, Roles.PHARMACIST, Roles.NURSE, Roles.ADMINISTRATOR})
-public class TriagePatientController extends Controller {
+@AllowedRoles({Roles.MANAGER})
+public class ManagerController extends Controller {
 
     private final ISessionService sessionService;
     private final IPatientService patientService;
@@ -41,15 +38,15 @@ public class TriagePatientController extends Controller {
     private final Form<CreateViewModel> createViewModelForm = Form.form(CreateViewModel.class);
 
     @Inject
-    public TriagePatientController(ISessionService sessionService, IPatientService patientService,
-                                   IEncounterService encounterService,ISearchService searchService) {
+    public ManagerController(ISessionService sessionService, IPatientService patientService,
+                             IEncounterService encounterService, ISearchService searchService) {
         this.sessionService = sessionService;
         this.patientService = patientService;
         this.encounterService=encounterService;
         this.searchService=searchService;
     }
 
-    public Result triagePatient() {
+    public Result indexGet() {
 
 
         CurrentUser currentUser = sessionService.retrieveCurrentUserSession();
@@ -66,11 +63,11 @@ public class TriagePatientController extends Controller {
         }
 
         //sets Patients Items in view model used
-        ManageViewModelPost viewModel = new ManageViewModelPost();
+        IndexViewModelGet viewModel = new IndexViewModelGet();
         viewModel.setTriagePatients(p);
 
 
-        return ok(triagePatient.render(currentUser, viewModel));
+        return ok(index.render(currentUser, viewModel));
     }
