+++ app/femr/ui/views/manager/index.scala.html
@@ -1,0 +1,41 @@
+
+@import femr.common.models.PatientItem
+@import femr.ui.controllers.manager.ManagerController
+@import femr.ui.views.html.layouts.main
+@import collection.JavaConversions._
+
+@(currentUser: femr.common.dtos.CurrentUser, ManageViewModelPost: femr.ui.models.manager.IndexViewModelGet)
+@additionalScripts = {
+        <script type="text/javascript" src="@routes.Assets.versioned("js/libraries/jquery-2.0.3.min.js")"></script>
+        <script type="text/javascript" src="@routes.Assets.versioned("js/libraries/jquery.tablescroll.js")"></script>
+        <script type="text/javascript" src="@routes.Assets.versioned("js/history/history.js")"></script>
+
+    }
+@additionalStyles = {
+        <link rel="stylesheet" href="@routes.Assets.versioned("css/libraries/jquery.tablescroll.css")">
+        <link rel="stylesheet" href="@routes.Assets.versioned("css/history.css")">
+    }
+
+<!doctype html>
+@main("Triage Patients Checked in Today", currentUser, scripts = additionalScripts, styles = additionalStyles){
+
+    <body>
+
+            <div id="center-panel">
+        <div class="backgroundForWrap">
+            <div  class ="h1" id="title"> List Of Triage Patients Checked In Today: @ManageViewModelPost.getToday</div>
+                <div  id="content-pane" >
+             <ol>
+                    @for(y <- 1 to ManageViewModelPost.getTriagePatients.size) {
+                        <li>@ManageViewModelPost.getTriagePatients(y - 1).getFirstName @ManageViewModelPost.getTriagePatients(y - 1).getLastName</li>
+
+                }
+               </ol>
+                </div>
+        </div>
+            </div>
+
+        </body>
+
+
+    }
