+++ conf/routes
@@ -29,9 +29,10 @@ POST        /admin/trips/edit/:id                              @femr.ui.controll
 GET         /admin/trips/teams                                 @femr.ui.controllers.admin.TripController.teamsGet()
 POST        /admin/trips/teams                                 @femr.ui.controllers.admin.TripController.teamsPost()
 GET         /admin/trips/cities                                @femr.ui.controllers.admin.TripController.citiesGet()
-GET         /admin/triagePatients/triagePatient                 @femr.ui.controllers.admin.TriagePatientController.triagePatient()
 POST        /admin/trips/cities                                @femr.ui.controllers.admin.TripController.citiesPost()
 GET         /admin                                             @femr.ui.controllers.admin.AdminController.index()
+#Manager
+GET         /manager                                           @femr.ui.controllers.manager.ManagerController.indexGet()
 #Pharmacy
 GET         /pharmacy/edit/:id                                 @femr.ui.controllers.PharmaciesController.editGet(id: Integer)
 POST        /pharmacy/edit/:id                                 @femr.ui.controllers.PharmaciesController.editPost(id: Integer)
